+++ hata/discord/interaction/application_command/application_command_option_choice.py
@@ -4,7 +4,7 @@ from enum import Enum
 
 from scarletio import RichAttributeErrorBaseType
 
-from ...localizations.helpers import localized_dictionary_builder
+from ...localizations.helpers import get_localized_length, localized_dictionary_builder
 from ...localizations.utils import build_locale_dictionary, destroy_locale_dictionary
 
 from .constants import (
@@ -269,14 +269,8 @@ class ApplicationCommandOptionChoice(RichAttributeErrorBaseType):
         """Returns the application command choice's length."""
         length = 0
         
-        # name
-        length += len(self.name)
-        
-        # name_localizations
-        name_localizations = self.name_localizations
-        if (name_localizations is not None):
-            for value in name_localizations.values():
-                length += len(value)
+        # name & name_localizations
+        length += get_localized_length(self.name, self.name_localizations)
         
         # value
         value = self.value