+++ hata/discord/oauth2/preinstanced.py
@@ -42,6 +42,8 @@ class ConnectionType(PreinstancedBase):
     +-----------------------+-----------------------+-------------------+
     | reddit                | Reddit                | reddit            |
     +-----------------------+-----------------------+-------------------+
+    | riot_games            | Riot Games            | riotgames         |
+    +-----------------------+-----------------------+-------------------+
     | samsung_galaxy        | Samsung Galaxy        | samsunggalaxy     |
     +-----------------------+-----------------------+-------------------+
     | spotify               | Spotify               | spotify           |
@@ -72,6 +74,7 @@ class ConnectionType(PreinstancedBase):
     league_of_legends = P('leagueoflegends', 'League of Legends')
     playstation = P('playstation', 'PlayStation Network')
     reddit = P('reddit', 'Reddit')
+    riot_games = P('riotgames', 'Riot Games')
     samsung_galaxy = P('samsunggalaxy', 'Samsung Galaxy')
     spotify = P('spotify', 'Spotify')
     skype = P('skype', 'Skype')